requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
aiohttp-socks==0.8.4
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
